IW-4071: extract logic to helper class
