+++ extensions/wikia/DiscussionModeration/DiscussionModerationController.class.php
@@ -1,10 +1,7 @@
 <?php
 
 use GuzzleHttp\Client;
-use GuzzleHttp\Psr7\Uri;
 use Wikia\Factory\ServiceFactory;
-use function GuzzleHttp\Psr7\build_query;
-use function GuzzleHttp\Psr7\parse_query;
 
 class DiscussionModerationController extends WikiaController {
 
@@ -15,6 +12,8 @@ class DiscussionModerationController extends WikiaController {
 	private $baseDomain;
 
 	private $scriptPath;
+
+	private $reportedPostsHelper;
 
 	public function __construct() {
 		parent::__construct();
@@ -29,6 +28,8 @@ class DiscussionModerationController extends WikiaController {
 
 		$this->baseDomain = $this->wg->Server;
 		$this->scriptPath = $this->wg->ScriptPath;
+
+		$this->reportedPostsHelper = new ReportedPostsHelper($this->baseDomain, $this->scriptPath );
 	}
 
 	public function init() {
@@ -49,72 +50,28 @@ class DiscussionModerationController extends WikiaController {
 
 		$isAnonOrBlockedUser = $this->isAnonOrBlockedUser( $user );
 
-		if ( $isAnonOrBlockedUser || !$user->isAllowed( 'read' ) || !$user->isAllowed( 'posts:validate' ) ) {
+		if ( $isAnonOrBlockedUser || !$user->isAllowed( 'read' ) ||
+			 !$user->isAllowed( 'posts:validate' ) ) {
 			$this->response->setCode( WikiaResponse::RESPONSE_CODE_FORBIDDEN );
+
 			return;
 		}
 
 		$canViewHidden = $user->isAllowed( 'posts:viewhidden' );
-		$canViewHiddenInContainer = $containerType == 'WALL' ? $user->isAllowed( 'wallremove' ) : $canViewHidden;
+		$canViewHiddenInContainer =
+			$containerType == 'WALL' ? $user->isAllowed( 'wallremove' ) : $canViewHidden;
 
-		$reportedPosts = $this->gateway->getReportedPosts(
-			$pagination, $viewableOnly, $canViewHidden, $canViewHiddenInContainer, $containerType, $user->getId()
-		);
+		$reportedPosts =
+			$this->gateway->getReportedPosts( $pagination, $viewableOnly, $canViewHidden,
+				$canViewHiddenInContainer, $containerType, $user->getId() );
 
 		if ( $reportedPosts ) {
-			$this->mapLinks( $reportedPosts );
-			$this->addPermissions( $user, $reportedPosts );
+			$this->reportedPostsHelper->mapLinks( $reportedPosts );
+			$this->reportedPostsHelper->addPermissions( $user, $reportedPosts );
 			$this->response->setData( $reportedPosts );
 			$this->response->setCode( WikiaResponse::RESPONSE_CODE_OK );
 		} else {
 			$this->response->setCode( WikiaResponse::RESPONSE_CODE_INTERNAL_SERVER_ERROR );
-		}
-	}
-
-	private function mapLinks( &$reportedPosts ) {
-		foreach ( $reportedPosts['_links'] as $linkName => $target ) {
-
-			if ( array_key_exists( 'href', $target ) ) {
-				$uri = new Uri( $target['href'] );
-				$reportedPosts['_links'][$linkName]['href'] = $this->buildNewLink( $uri );
-			} else {
-				$uri = new Uri( $target[0]['href'] );
-				$reportedPosts['_links'][$linkName][0]['href'] = $this->buildNewLink( $uri );
-			}
-		}
-	}
-
-	private function buildNewLink( Uri $uri ) {
-		$serviceQueryParams = parse_query( $uri->getQuery() );
-		$controllerQueryParams = [
-			'controller' => 'DiscussionModerationController',
-			'method' => 'getReportedPosts',
-		];
-
-		foreach ( $serviceQueryParams as $paramName => $value ) {
-			$controllerQueryParams[$paramName] = $value;
-		}
-		return (string)$uri->withScheme( 'https' )
-			->withHost( $this->baseDomain )
-			->withPath( $this->scriptPath . '/wikia.php' )
-			->withQuery( build_query( $controllerQueryParams ) );
-	}
-
-	private function addPermissions( User $user, array &$reportedPosts ) {
-		foreach ( $reportedPosts['_embedded']['doc:posts'] as &$postData ) {
-			$post = ( new PostBuilder() )
-				->position( $postData['position'] )
-				->author( empty( $postData['creatorId'] ) ? 0 : (int)$postData['creatorId'] )
-				->creationDate( $postData['creationDate']['epochSecond'] )
-				->isEditable( $postData['isEditable'] )
-				->isThreadEditable( $postData['_embedded']['thread'][0]['isEditable'] )
-				->build();
-
-			$rights = DiscussionPermissionManager::getRights( $user, $post );
-
-			if ( !empty( $rights ) ) {
-				$postData['_embedded']['userData'][0]['permissions'] = $rights;
-			}
 		}
 	}
